imrprove readme
