+++ src/main/scala/v1/div/properties/ModIdempotence.scala
@@ -76,87 +76,4 @@ object ModIdempotence {
 
     Div(a, b, divx, modx).solve == Div(a, b, divy, mody).solve
   }.holds
-
-  def modAdd(a: BigInt, b: BigInt, c: BigInt): Boolean = {
-    require(a >= 0)
-    require(c >= 0)
-    require(b > 0)
-
-    val x = Div(a, b, 0, a)
-    val solvedX = x.solve
-    check(solvedX.isFinal)
-    check(solvedX.isValid)
-    check(solvedX.mod < solvedX.b)
-    check(solvedX.mod < b)
-    check(solvedX.a == a)
-    check(solvedX.b == b)
-    check(solvedX.a == solvedX.b * solvedX.div + solvedX.mod)
-    check(solvedX.a - solvedX.b * solvedX.div == solvedX.mod)
-
-    val y = Div(c, b, 0, c)
-    val solvedY = y.solve
-    check(solvedY.isFinal)
-    check(solvedY.isValid)
-    check(solvedY.mod < solvedY.b)
-    check(solvedY.mod < b)
-    check(solvedY.a == c)
-    check(solvedY.b == b)
-    check(solvedY.a == solvedY.b * solvedY.div + solvedY.mod)
-    check(solvedY.a - solvedY.b * solvedY.div == solvedY.mod)
-
-    val xy = Div(a + c, b, 0, a + c)
-    val solvedXY = xy.solve
-    check(solvedXY.isFinal)
-    check(solvedXY.isValid)
-    check(solvedXY.mod < solvedXY.b)
-    check(solvedXY.mod < b)
-    check(solvedXY.a == a + c)
-    check(solvedXY.b == b)
-    check(solvedXY.a == solvedXY.b * solvedXY.div + solvedXY.mod)
-    check(a + c == b * solvedXY.div + solvedXY.mod)
-
-    val z = Div( solvedX.mod + solvedY.mod, b, 0, solvedX.mod + solvedY.mod)
-    check(z.a == z.b * z.div + z.mod)
-    check(z.a == solvedX.mod + solvedY.mod)
-    check(z.b == b)
-    check(z.mod == solvedX.mod + solvedY.mod)
-    check(z.div == 0)
-    check(z.a == z.b * z.div + z.mod)
-
-    val solvedZ = z.solve
-    check(solvedZ.isValid && solvedZ.isFinal)
-    check(solvedZ.mod < z.b)
-    check(modUniqueDiv(z, solvedZ))
-    check(z.solve.mod == solvedZ.mod)
-
-    check(solvedX.mod + solvedY.mod == b * solvedZ.div + solvedZ.mod)
-    check(solvedX.a - solvedX.b * solvedX.div + solvedY.a - solvedY.b * solvedY.div == b * solvedZ.div + solvedZ.mod)
-    check(a - b * solvedX.div + c - b * solvedY.div == b * solvedZ.div + solvedZ.mod)
-    check(a + c == b * solvedZ.div + b * solvedX.div + b * solvedY.div + solvedZ.mod)
-
-    val bigDiv = solvedZ.div + solvedX.div + solvedY.div
-    check(a + c == b * bigDiv + solvedZ.mod)
-
-    val w = Div(a + c, b, bigDiv, solvedZ.mod)
-    check(solvedZ.mod < b)
-    check(w.mod == solvedZ.mod)
-    check(w.isFinal)
-    check(w.solve == w)
-
-    check(bigDiv >= 0)
-    DivModAdditionAndMultiplication.APlusMultipleTimesBSameMod(a + c, b, bigDiv)
-    check( Calc.mod(a + c,b) == Calc.mod( a + c + b * bigDiv, b ))
-    check(w.isValid)
-    check(xy.isValid)
-    check(w.a == xy.a)
-    check(w.b == xy.b)
-    check(w.b > 0) // @todo check that also works for b < 0
-    check(modUniqueDiv(w, xy))
-    check( w.solve == xy.solve)
-    check( w.solve.mod == xy.solve.mod)
-    check( xy.solve.mod == Calc.mod(a+c,b))
-    check( xy.solve.mod == solvedZ.mod)
-
-    Calc.mod(a + c, b) == Calc.mod(Calc.mod(a, b) + Calc.mod(c, b), b)
-  }.holds
 }
